Organize folder structure for rsa
